move notifications actions to settings actions
